# Phase 1 MVP: Interactive Map, Timeline Visualization & Filtering System