--- a/src/App.tsx
+++ b/src/App.tsx
@@ -1,0 +1,39 @@
+function App() {
+  return (
+    <div className="min-h-screen bg-gray-50">
+      {/* Header */}
+      <header className="bg-gray-900 text-white shadow-lg">
+        <div className="container mx-auto px-4 py-6">
+          <h1 className="text-3xl font-bold">Heritage Tracker</h1>
+          <p className="text-gray-300 mt-2">
+            Documenting the destruction of cultural heritage in Gaza (2023-2024)
+          </p>
+        </div>
+      </header>
+
+      {/* Main Content */}
+      <main className="container mx-auto px-4 py-8">
+        <div className="bg-white rounded-lg shadow-md p-8 text-center">
+          <h2 className="text-2xl font-semibold text-gray-800 mb-4">
+            Interactive Map Coming Soon
+          </h2>
+          <p className="text-gray-600">
+            Building the application incrementally. Dev server is running with HMR enabled.
+          </p>
+        </div>
+      </main>
+
+      {/* Footer */}
+      <footer className="bg-gray-900 text-gray-400 mt-16">
+        <div className="container mx-auto px-4 py-6 text-center">
+          <p className="text-sm">
+            Heritage Tracker • Evidence-based documentation • All data verified by UNESCO,
+            Forensic Architecture, and Heritage for Peace
+          </p>
+        </div>
+      </footer>
+    </div>
+  );
+}
+
+export default App;