--- a/src/App.tsx
+++ b/src/App.tsx
@@ -1,5 +1,6 @@
 import { mockSites } from "./data/mockSites";
-import { components, getStatusColor, cn } from "./styles/theme";
+import { components, cn } from "./styles/theme";
+import { SiteCard } from "./components/SiteCard";
 
 function App() {
   return (
@@ -30,48 +31,7 @@ function App() {
         {/* Sites List */}
         <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
           {mockSites.map((site) => (
-            <div key={site.id} className={components.card.base}>
-              {/* Status Badge */}
-              <div className={cn(getStatusColor(site.status), "px-4 py-2 text-sm font-semibold text-white")}>
-                {site.status.toUpperCase().replace("-", " ")}
-              </div>
-
-              {/* Content */}
-              <div className={components.card.padding}>
-                <h3 className="text-xl font-bold text-gray-900 mb-2">{site.name}</h3>
-                {site.nameArabic && (
-                  <p className="text-gray-600 text-sm mb-3">{site.nameArabic}</p>
-                )}
-
-                <div className="space-y-2 text-sm text-gray-600 mb-4">
-                  <p>
-                    <span className="font-semibold">Type:</span> {site.type.replace("-", " ")}
-                  </p>
-                  <p>
-                    <span className="font-semibold">Built:</span> {site.yearBuilt}
-                  </p>
-                  {site.dateDestroyed && (
-                    <p>
-                      <span className="font-semibold">Destroyed:</span> {site.dateDestroyed}
-                    </p>
-                  )}
-                </div>
-
-                <p className="text-gray-700 text-sm line-clamp-3">{site.description}</p>
-
-                {/* Verified By */}
-                <div className="mt-4 pt-4 border-t border-gray-200">
-                  <p className="text-xs text-gray-500">Verified by:</p>
-                  <div className="flex flex-wrap gap-2 mt-2">
-                    {site.verifiedBy.map((org) => (
-                      <span key={org} className={components.badge.primary}>
-                        {org}
-                      </span>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            </div>
+            <SiteCard key={site.id} site={site} />
           ))}
         </div>
       </main>