## Summary

Complete Phase 1 MVP implementation for Heritage Tracker - an interactive web application documenting Palestinian cultural heritage destruction in Gaza (2023-2024).

This PR includes 5 major development phases with full testing, documentation, and code quality checks.

---

## 🎯 What's Included

### Phase A: Project Setup
- ✅ React 19 + TypeScript + Vite project initialization
- ✅ Tailwind CSS v4 configuration with @tailwindcss/postcss
- ✅ ESLint + Prettier setup
- ✅ Clean header/footer layout

### Phase B: Data Architecture  
- ✅ TypeScript interfaces (GazaSite, Source)
- ✅ Centralized theme system (src/styles/theme.ts)
- ✅ 3 initial mock heritage sites with full data
- ✅ Color-coded status badges

### Phase C: Component Abstraction
- ✅ Extracted reusable SiteCard component
- ✅ RTL support for Arabic text
- ✅ 43% reduction in App.tsx complexity

### Phase D: Interactive Map
- ✅ Leaflet + React-Leaflet integration
- ✅ Color-coded markers (red/orange/yellow by status)
- ✅ Browser language detection (English/Arabic tiles)
- ✅ Scrollable popups with site details
- ✅ Vitest + React Testing Library setup

### Phase E: Timeline & Filtering System
- ✅ D3.js interactive timeline visualization
- ✅ Smart tooltip positioning (edge-aware)
- ✅ Date-based filtering ("on or before" logic)
- ✅ Type and Status dropdown filters
- ✅ Synchronized filter state across all views
- ✅ Enhanced theme system (form components)

---

## 📊 Current Stats

- **Heritage Sites:** 5 sites spanning Oct 2023 - Feb 2024
- **Test Coverage:** 7 passing tests (~6 seconds)
- **Code Quality:** All ESLint checks passing, TypeScript strict mode
- **Files Changed:** 14 files, 1,251+ insertions
- **Performance:** Optimized Vitest configuration (8x speedup)

---

## ✨ Key Features

### Interactive Map
- Display heritage sites on Gaza-focused map
- Color-coded markers by damage status
- Click markers for detailed popups
- Browser language detection for map tiles

### Timeline Visualization
- D3.js horizontal timeline (Oct 2023 - Feb 2024)
- Click markers to filter by date
- Visual feedback for selected dates
- Smart tooltips that adjust position at edges

### Comprehensive Filtering
- **Timeline:** Filter by destruction date
- **Type:** mosque, church, archaeological, museum, historic-building
- **Status:** destroyed, heavily-damaged, damaged
- All filters combine with AND logic
- Dynamic site count ("X of Y")

### Responsive Design
- Mobile-first approach
- RTL support for Arabic
- Tailwind CSS v4 with centralized theme
- Consistent styling across components

---

## 🧪 Testing

All tests passing:
```bash
npm test -- --run
```

- ✅ App component (integration test)
- ✅ SiteCard component (2 tests)
- ✅ Timeline component (2 tests)  
- ✅ Filters component (2 tests)
- ✅ HeritageMap component (smoke test)

**Performance:** Tests run in ~6 seconds (optimized Vitest config)

---

## 📝 Code Quality

- ✅ ESLint passing (no errors or warnings)
- ✅ TypeScript strict mode enabled
- ✅ Code reviewed for DRY/KISS/SOLID principles
- ✅ Centralized theme system prevents duplication
- ✅ Proper TypeScript types (no `any`)

---

## 🗂️ Project Structure

```
src/
├── components/
│   ├── Filters/          # Type and Status filter dropdowns
│   ├── Map/              # Leaflet map with markers
│   ├── Timeline/         # D3.js timeline visualization
│   └── SiteCard.tsx      # Reusable site display card
├── data/
│   └── mockSites.ts      # 5 heritage sites (Oct 2023 - Feb 2024)
├── styles/
│   └── theme.ts          # Centralized theme system
├── types/
│   └── index.ts          # TypeScript interfaces
└── test/
    └── setup.ts          # Vitest configuration
```

---

## 🚀 Next Steps (Post-Merge)

Based on CURRENT_SESSION.md priorities:

1. **Data Collection** - Add 15-20 more heritage sites
2. **Detail Panel/Modal** - In-depth site information view
3. **Timeline Animation** - Play button for progression
4. **Search Functionality** - Filter by site name
5. **Statistics Dashboard** - Landing page with impact numbers

---

## 📚 Documentation

- ✅ Comprehensive CURRENT_SESSION.md with phase summaries
- ✅ CLAUDE.md with lessons learned and technical decisions
- ✅ Inline code comments and JSDoc
- ✅ Test files with clear descriptions

---

## 🔍 Review Notes

### Key Technical Decisions

**Leaflet vs Mapbox:** Chose Leaflet (no API key, free tiles, simpler)  
**Vitest vs Jest:** Chose Vitest (10x faster, native Vite integration)  
**D3.js vs Pure SVG:** Chose D3.js (elegant scales/axes, worth bundle size)  
**Incremental Development:** Built features piece by piece, verified at each step

### Challenges Solved

1. **Coordinate Format:** Fixed [lng,lat] → [lat,lng] for Leaflet
2. **Tooltip Overflow:** Smart positioning at timeline edges
3. **Selected State:** Event handler state management for markers
4. **Test Performance:** Optimized from 45s → 6s with Vitest config
5. **Arabic Map Tiles:** Browser language detection for i18n

---

## ✅ Checklist

- [x] All tests passing
- [x] ESLint checks passing
- [x] TypeScript strict mode enabled
- [x] Documentation updated
- [x] Code reviewed for best practices
- [x] Dev server running without errors
- [x] Responsive design verified
- [x] Git history clean (10 commits)

---

## 🎥 Demo

The application is fully functional at `http://localhost:5174` with:
- 5 heritage sites displayed on interactive map
- Color-coded markers and timeline
- Working filters (date, type, status)
- Synchronized updates across all views

---

🤖 Generated with [Claude Code](https://claude.com/claude-code)